% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colRanks_fast}
\alias{colRanks_fast}
\title{Computes rankings for each column of a matrix in parallel.}
\usage{
colRanks_fast(x, n_cores)
}
\arguments{
\item{x}{The input matrix to be ranked.}

\item{n_cores}{The number of CPU cores to use for parallel processing.}
}
\value{
A matrix where each column contains the rankings for the
corresponding column in the input matrix.
}
\description{
Computes rankings for each column of a matrix in parallel.
}
