% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cor_fast}
\alias{cor_fast}
\title{Computes a Pearson}
\usage{
cor_fast(x, n_cores)
}
\arguments{
\item{x}{The input matrix, where each column represents a set of
observations.}

\item{n_cores}{The number of CPU cores to utilize for parallel computation
(optional, defaults to 1).}
}
\value{
A Pearson correlation matrix if `useRanks` is `false`. If
`useRanks` is `true`, returns a Spearman correlation matrix.
}
\description{
This function calculates a Pearson correlation matrix
}
