% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{getPConsMtx}
\alias{getPConsMtx}
\title{Get processed consensus matrix.}
\usage{
getPConsMtx(kmResults, consMin)
}
\arguments{
\item{kmResults}{list of k-means clustering assignments on the PCA loadings
sub-datasets.}

\item{consMin}{the low-pass filter threshold value for processed
consensus matrix.}
}
\value{
a processed consensus matrix.
}
\description{
This function gets consensus matrix based on the clustering solutions
contained in the \code{kmResults} input parameter and does the processing to
use it for imputation.
}
