% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getScale}
\alias{getScale}
\title{Computes Means and Standard Deviations for Scaling}
\usage{
getScale(x, n_cores)
}
\arguments{
\item{x}{A numeric matrix representing the gene expression data, where rows
are genes and columns are samples.}

\item{n_cores}{The number of cores to use for parallel processing.}
}
\value{
A list containing:
  * `means`: A numeric vector of column means.
  * `sds`: A numeric vector of column standard deviations.
}
\description{
Computes Means and Standard Deviations for Scaling
}
