% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{impute}
\alias{impute}
\title{Performs imputation of dropout values in scRNA-seq data using ccImpute
algorithm}
\usage{
impute(
  logX,
  useRanks = TRUE,
  pcaMin,
  pcaMax,
  k,
  consMin = 0.65,
  nCores,
  kmNStart,
  kmMax = 1000
)
}
\arguments{
\item{logX}{A normalized and log transformed scRNA-seq expression matrix.}

\item{useRanks}{A Boolean specifying if non-parametric version of weighted
Pearson correlation should be used.}

\item{pcaMin}{This is used to establish the number of minimum PCA features
used for generating subsets. For small datasets up to \code{500} cells this
equals pcaMin*n minimum features, where \code{n} is number of cells. For
large datasets, this corresponds to the feature count that has proportion of
variance less than \code{pcaMin}. Both pcaMin and pcaMax must be specified
to be considered.}

\item{pcaMax}{This is used to establish the number of maximum PCA features
used for generating subsets. For small datasets up to \code{500} cells this
equals pcaMax*n maximum features, where \code{n} is number of cells. For
large datasets, this corresponds to the feature count that has proportion of
variance less than \code{pcaMax}. Both pcaMin and pcaMax must be specified
to be considered.}

\item{k}{centers parameter passed to \code{\link[stats]{kmeans}} function}

\item{consMin}{the low-pass filter threshold for processing consensus
matrix.}

\item{nCores}{the number of cores to be used on the user's machine.
If not set, \code{ccImpute} will use all but one cores of your machine.}

\item{kmNStart}{nstart parameter passed to \code{\link[stats]{kmeans}}.
function. Can be set manually. By default it is \code{1000} for up to
\code{2000} cells and \code{50} for more than \code{2000} cells.}

\item{kmMax}{iter.max parameter passed to \code{\link[stats]{kmeans}}.}
}
\value{
A normalized and log transformed scRNA-seq expression matrix with
imputed missing values.
}
\description{
Performs imputation of dropout values in scRNA-seq data using ccImpute
algorithm
}
