% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{kmeans}
\alias{kmeans}
\title{This function performs \code{\link[stats]{kmeans}} clustering of the
subdatasets corresponding to varying number of PCA loadings as contained in
\code{nDim} input parameter.}
\usage{
kmeans(input, k, nCores, nDim, kmNStart, kmMax)
}
\arguments{
\item{input}{the matrix of all variable loadings.}

\item{k}{centers (integer) parameter passed to \code{\link[stats]{kmeans}}
function.}

\item{nCores}{defines the number of cores to be used on the user's machine.
If not set, `ccImpute` will use all but one cores of your machine.}

\item{nDim}{the list of containing a number of PCA loadings to use for each
sub-dataset.}

\item{kmNStart}{nstart parameter passed to \code{\link[stats]{kmeans}}
function. Can be set manually. By default it is \code{1000} for up to
\code{2000} cells and \code{50} for more than \code{2000} cells.}

\item{kmMax}{iter.max parameter passed to
\code{\link[stats]{kmeans}} function.}
}
\value{
a list of clustering assignments for all the sub-datasets.
}
\description{
This function performs \code{\link[stats]{kmeans}} clustering of the
subdatasets corresponding to varying number of PCA loadings as contained in
\code{nDim} input parameter.
}
