% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{printer}
\alias{printer}
\title{Internal Printing Utility}
\usage{
printer(verbose, msg, startTime)
}
\arguments{
\item{verbose}{logical. If \code{TRUE}, messages are printed to the console. 
If \code{FALSE}, messages are suppressed.}

\item{msg}{The message to be printed.}

\item{startTime}{A timestamp indicating the start time for 
elapsed time calculation. If omitted, no elapsed time is shown.}
}
\value{
No return
}
\description{
This internal function provides a flexible way to print messages
to the console, optionally including elapsed time information.
}
\keyword{internal}
