% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solveDrops}
\alias{solveDrops}
\title{Computes imputed expression matrix using linear eq solver.}
\usage{
solveDrops(cm, em, ids, n_cores)
}
\arguments{
\item{cm}{processed consensus matrix}

\item{em}{expression matrix}

\item{ids}{location of values determined to be dropout events}

\item{n_cores}{number of cores to use for parallel computation.}
}
\value{
imputed expression matrix
}
\description{
Computes imputed expression matrix using linear eq solver.
}
