% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solver2}
\alias{solver2}
\title{Fast Calculation of "Dropout" values}
\usage{
solver2(cm, em, ids, n_cores)
}
\arguments{
\item{cm}{A numeric matrix representing the consensus matrix.}

\item{em}{A dense numeric matrix representing the gene expression data,
where rows are genes and columns are samples.}

\item{ids}{An integer matrix specifying the row and column indices of
entries for which to calculate importance scores. Each row of `ids` should
contain two integers: the row index (gene) and column index (sample) in the
`em` matrix.}

\item{n_cores}{The number of cores to use for parallel processing.}
}
\value{
A numeric vector of imputed dropout values, corresponding to the
entries specified in the `ids` matrix.
}
\description{
Fast Calculation of "Dropout" values
}
