% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wCor_fast}
\alias{wCor_fast}
\title{Computes a Weighted Pearson}
\usage{
wCor_fast(x, w, n_cores)
}
\arguments{
\item{x}{The input matrix (dense), where each column represents a set of
observations.}

\item{w}{A vector of weights, one for each observation (must have the same
number of elements as rows in `x`).}

\item{n_cores}{The number of CPU cores to utilize for parallel computation.}
}
\value{
A weighted Pearson correlation matrix
}
\description{
This function calculates weighted Pearson correlation matrix
}
