% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wCorDist}
\alias{wCorDist}
\title{Computes a weighted Pearson distance measure matrix. If ranks are used
this measure turns into weighted Spearman distance measure matrix.}
\usage{
wCorDist(x, w, useRanks, n_cores)
}
\arguments{
\item{x}{input with columns containing each observation}

\item{w}{weights for all values in a obervation}

\item{useRanks}{indicates if Pearson should be computed on weighted ranks.}

\item{n_cores}{number of cores to use for parallel computation.}
}
\value{
weighted Pearson distance measure matrix. If ranks are used
this measure turns into weighted Spearman distance measure matrix.
}
\description{
Computes a weighted Pearson distance measure matrix. If ranks are used
this measure turns into weighted Spearman distance measure matrix.
}
