% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{ccImpute}
\alias{ccImpute}
\title{Performs imputation of dropout values in scRNA-seq data using ccImpute
algorithm as described in the ccImpute: an accurate and scalable consensus 
clustering based algorithm to impute dropout events in the single-cell 
RNA-seq data DOI: https://doi.org/10.1186/s12859-022-04814-8}
\usage{
ccImpute(
    logX,
    useRanks = TRUE,
    pcaMin,
    pcaMax,
    k,
    consMin = 0.65,
    kmNStart,
    kmMax = 1000
)
}
\arguments{
\item{logX}{A normalized and log transformed scRNA-seq expression matrix.}

\item{useRanks}{A Boolean specifying if non-parametric version of weighted
Pearson correlation should be used. It's recommended to keep this as TRUE
since  this performs better as determined experimentally. However, FALSE
will also provide decent results with the benefit or faster runtime.}

\item{pcaMin}{This is used to establish the number of minimum PCA features
used for generating subsets. For small datasets up to \code{500} cells this
equals pcaMin*n minimum features, where \code{n} is number of cells. For
large datasets, this corresponds to the feature count that has proportion of
variance less than \code{pcaMin}. Both pcaMin and pcaMax must be specified
to be considered. It's best to keep this value as default unless a better
value was obtained experimentally.}

\item{pcaMax}{This is used to establish the number of maximum PCA features
used for generating subsets. For small datasets up to \code{500} cells this
equals pcaMax*n maximum features, where \code{n} is number of cells. For
large datasets, this corresponds to the feature count that has proportion of
variance less than \code{pcaMax}. Both pcaMin and pcaMax must be specified
to be considered. It's best to keep this value as default unless a better
value was obtained experimentally.}

\item{k}{centers parameter passed to \code{\link[stats]{kmeans}} function.
This corresponds to a number of different cell groups in data. This can be
estimated in a number of methods. If not provided we take the approach
provided in the SIMLR package. 
(https://www.bioconductor.org/packages/release/bioc/html/SIMLR.html)}

\item{consMin}{the low-pass filter threshold for processing consensus
matrix. This is to eliminate noise from unlikely clustering assignmnets.
It is recommended to keep this value >-.5.}

\item{kmNStart}{nstart parameter passed to \code{\link[stats]{kmeans}}.
function. Can be set manually. By default it is \code{1000} for up to
\code{2000} cells and \code{50} for more than \code{2000} cells.}

\item{kmMax}{iter.max parameter passed to \code{\link[stats]{kmeans}}.
\code{ccImpute} is a stochastic method, and setting the \code{rand_seed}
allows reproducibility.}
}
\value{
A normalized and log transformed scRNA-seq expression matrix with
imputed missing values.
}
\description{
Performs imputation of dropout values in scRNA-seq data using ccImpute
algorithm as described in the ccImpute: an accurate and scalable consensus 
clustering based algorithm to impute dropout events in the single-cell 
RNA-seq data DOI: https://doi.org/10.1186/s12859-022-04814-8
}
\examples{
exp_matrix <- log(abs(matrix(rnorm(1000000),nrow=10000))+1)
ccImpute(exp_matrix, k = 2)
}
