% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getConsMtx}
\alias{getConsMtx}
\title{Computes consensus matrix given cluster labels}
\usage{
getConsMtx(dat)
}
\arguments{
\item{dat}{a matrix containing clustering solutions in columns}
}
\value{
consensus matrix
}
\description{
Computes consensus matrix given cluster labels
}
